Created new subpackage for enum files and moved the existing files there
