+++ server/backend/src/main/java/com/cropify/dto/AgricultureProductsDTO.java
@@ -1,7 +1,7 @@
 package com.cropify.dto;
 
 
-import com.cropify.entity.AgriProductType;
+import com.cropify.entity.enums.AgriProductType;
 
 public class AgricultureProductsDTO {
