+++ server/backend/src/main/java/com/cropify/dto/FarmProductsDTO.java
@@ -1,6 +1,6 @@
 package com.cropify.dto;
 
-import com.cropify.entity.FarmProductType;
+import com.cropify.entity.enums.FarmProductType;
 
 public class FarmProductsDTO {
