+++ server/backend/src/main/java/com/cropify/entity/UserDetails.java
@@ -15,6 +15,8 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
 
+import com.cropify.entity.enums.UserType;
+
 
 @Entity
 public class UserDetails {
