+++ server/backend/src/main/java/com/cropify/entity/enums/FarmProductType.java
@@ -1,4 +1,4 @@
-package com.cropify.entity;
+package com.cropify.entity.enums;
 
 public enum FarmProductType {
 	P,Q,R,S
