+++ server/backend/src/main/java/com/cropify/entity/enums/FarmProductsStatus.java
@@ -1,4 +1,4 @@
-package com.cropify.entity;
+package com.cropify.entity.enums;
 
 public enum FarmProductsStatus {
 	FRESH,MATURE,EXPIRED
