+++ server/backend/src/main/java/com/cropify/entity/enums/MachineType.java
@@ -1,4 +1,4 @@
-package com.cropify.entity;
+package com.cropify.entity.enums;
 
 public enum MachineType {
 	A,B,C,D
